+++ deltachat-ffi/deltachat.h
@@ -3431,25 +3431,24 @@ dc_lot_t*        dc_chatlist_get_summary     (const dc_chatlist_t* chatlist, siz
 
 
 /**
- * Create a chatlist summary item when the chatlist object is already unref()'d.
+ * Create a chat summary item.
  *
- * This function is similar to dc_chatlist_get_summary(), however,
- * it takes the chat ID and the message ID as returned by dc_chatlist_get_chat_id() and dc_chatlist_get_msg_id()
- * as arguments. The chatlist object itself is not needed directly.
+ * This function is similar to dc_chatlist_get_summary(), however, it takes the chat ID as returned
+ * by dc_chatlist_get_chat_id(), as an argument. But note that as this function does not use a
+ * chatlist object, it also does not use message IDs cached in the chatlist and returns a summary
+ * for the actual last message in the chat. So, consequent calls to this function may produce
+ * different summaries.
  *
- * This maybe useful if you convert the complete object into a different representation
- * as done e.g. in the node-bindings.
- * If you have access to the chatlist object in some way, using this function is not recommended,
- * use dc_chatlist_get_summary() in this case instead.
+ * This function also may be useful if you convert the complete object into a different
+ * representation as done e.g. in the node-bindings.
  *
  * @memberof dc_context_t
  * @param context The context object.
  * @param chat_id The chat ID to get a summary for.
- * @param msg_id The message ID to get a summary for.
  * @return The summary as an dc_lot_t object, see dc_chatlist_get_summary() for details.
  *     Must be freed using dc_lot_unref(). NULL is never returned.
  */
-dc_lot_t*        dc_chatlist_get_summary2    (dc_context_t* context, uint32_t chat_id, uint32_t msg_id);
+dc_lot_t*        dc_chat_get_summary         (dc_context_t* context, uint32_t chat_id);
 
 
 /**