+++ examples/simple.rs
@@ -86,8 +86,9 @@ async fn main() {
     let chats = Chatlist::try_load(&ctx, 0, None, None).await.unwrap();
 
     for i in 0..chats.len() {
-        let msg = Message::load_from_db(&ctx, chats.get_msg_id(i).unwrap().unwrap())
+        let msg = Message::load_from_db_last_for_chat(&ctx, chats.get_chat_id(i).unwrap())
             .await
+            .unwrap()
             .unwrap();
         log::info!("[{}] msg: {:?}", i, msg);
     }