+++ node/src/module.c
@@ -1804,14 +1804,13 @@ NAPI_METHOD(dcn_chatlist_get_summary) {
   return result;
 }
 
-NAPI_METHOD(dcn_chatlist_get_summary2) {
+NAPI_METHOD(dcn_chat_get_summary) {
   NAPI_ARGV(3);
   NAPI_DCN_CONTEXT();
   NAPI_ARGV_INT32(chat_id, 1);
-  NAPI_ARGV_INT32(message_id, 2);
 
   //TRACE("calling..");
-  dc_lot_t* summary = dc_chatlist_get_summary2(dcn_context->dc_context, chat_id, message_id);
+  dc_lot_t* summary = dc_chat_get_summary(dcn_context->dc_context, chat_id);
 
   napi_value result;
   if (summary == NULL) {
@@ -3533,7 +3532,7 @@ NAPI_INIT() {
   NAPI_EXPORT_FUNCTION(dcn_chatlist_get_cnt);
   NAPI_EXPORT_FUNCTION(dcn_chatlist_get_msg_id);
   NAPI_EXPORT_FUNCTION(dcn_chatlist_get_summary);
-  NAPI_EXPORT_FUNCTION(dcn_chatlist_get_summary2);
+  NAPI_EXPORT_FUNCTION(dcn_chat_get_summary);
 
   /**
    * dc_contact_t